+++ static/js/markdown.js
@@ -7,6 +7,7 @@ import * as fenced_code from "../shared/js/fenced_code";
 import marked from "../third/marked/lib/marked";
 
 import * as blueslip from "./blueslip";
+import * as linkifiers from "./linkifiers";
 import * as message_store from "./message_store";
 
 // This contains zulip's frontend Markdown implementation; see
@@ -22,9 +23,6 @@ import * as message_store from "./message_store";
 // See the call to markdown.initialize() in ui_init
 // for example usage.
 let helpers;
-
-const linkifier_map = new Map();
-let linkifier_list = [];
 
 // Regexes that match some of our common backend-only Markdown syntax
 const backend_only_markdown_re = [
@@ -88,6 +86,7 @@ export function contains_backend_only_syntax(content) {
     // If a linkifier doesn't start with some specified characters
     // then don't render it locally. It is workaround for the fact that
     // javascript regex doesn't support lookbehind.
+    const linkifier_list = linkifiers.linkifier_list;
     const false_linkifier_match = linkifier_list.find((re) => {
         const pattern = /[^\s"'(,:<]/.source + re.pattern.source + /(?!\w)/.source;
         const regex = new RegExp(pattern);
@@ -223,6 +222,7 @@ export function add_topic_links(message) {
     }
     const topic = message.topic;
     const links = [];
+    const linkifier_list = linkifiers.linkifier_list;
 
     for (const linkifier of linkifier_list) {
         const pattern = linkifier.pattern;
@@ -359,20 +359,6 @@ function handleStreamTopic(stream_name, topic) {
     )}" href="/${_.escape(href)}">${_.escape(text)}</a>`;
 }
 
-function handleLinkifier(pattern, matches) {
-    let url = linkifier_map.get(pattern);
-
-    let current_group = 1;
-
-    for (const match of matches) {
-        const back_ref = "\\" + current_group;
-        url = url.replace(back_ref, match);
-        current_group += 1;
-    }
-
-    return url;
-}
-
 function handleTex(tex, fullmatch) {
     try {
         return katex.renderToString(tex);
@@ -386,90 +372,7 @@ function handleTex(tex, fullmatch) {
     }
 }
 
-function python_to_js_linkifier(pattern, url) {
-    // Converts a python named-group regex to a javascript-compatible numbered
-    // group regex... with a regex!
-    const named_group_re = /\(?P<([^>]+?)>/g;
-    let match = named_group_re.exec(pattern);
-    let current_group = 1;
-    while (match) {
-        const name = match[1];
-        // Replace named group with regular matching group
-        pattern = pattern.replace("(?P<" + name + ">", "(");
-        // Replace named reference in URL to numbered reference
-        url = url.replace("%(" + name + ")s", "\\" + current_group);
-
-        // Reset the RegExp state
-        named_group_re.lastIndex = 0;
-        match = named_group_re.exec(pattern);
-
-        current_group += 1;
-    }
-    // Convert any python in-regex flags to RegExp flags
-    let js_flags = "g";
-    const inline_flag_re = /\(\?([Limsux]+)\)/;
-    match = inline_flag_re.exec(pattern);
-
-    // JS regexes only support i (case insensitivity) and m (multiline)
-    // flags, so keep those and ignore the rest
-    if (match) {
-        const py_flags = match[1].split("");
-
-        for (const flag of py_flags) {
-            if ("im".includes(flag)) {
-                js_flags += flag;
-            }
-        }
-
-        pattern = pattern.replace(inline_flag_re, "");
-    }
-    // Ideally we should have been checking that linkifiers
-    // begin with certain characters but since there is no
-    // support for negative lookbehind in javascript, we check
-    // for this condition in `contains_backend_only_syntax()`
-    // function. If the condition is satisfied then the message
-    // is rendered locally, otherwise, we return false there and
-    // message is rendered on the backend which has proper support
-    // for negative lookbehind.
-    pattern = pattern + /(?!\w)/.source;
-    let final_regex = null;
-    try {
-        final_regex = new RegExp(pattern, js_flags);
-    } catch (error) {
-        // We have an error computing the generated regex syntax.
-        // We'll ignore this linkifier for now, but log this
-        // failure for debugging later.
-        blueslip.error("python_to_js_linkifier: " + error.message);
-    }
-    return [final_regex, url];
-}
-
-export function update_linkifier_rules(linkifiers) {
-    // Update the marked parser with our particular set of linkifiers
-    linkifier_map.clear();
-    linkifier_list = [];
-
-    const marked_rules = [];
-
-    for (const linkifier of linkifiers) {
-        const [regex, final_url] = python_to_js_linkifier(linkifier.pattern, linkifier.url_format);
-        if (!regex) {
-            // Skip any linkifiers that could not be converted
-            continue;
-        }
-
-        linkifier_map.set(regex, final_url);
-        linkifier_list.push({
-            pattern: regex,
-            url_format: final_url,
-        });
-        marked_rules.push(regex);
-    }
-
-    marked.InlineLexer.rules.zulip.linkifiers = marked_rules;
-}
-
-export function initialize(linkifiers, helper_config) {
+export function initialize(helper_config) {
     helpers = helper_config;
 
     function disable_markdown_regex(rules, name) {
@@ -528,8 +431,6 @@ export function initialize(linkifiers, helper_config) {
     // Disable autolink as (a) it is not used in our backend and (b) it interferes with @mentions
     disable_markdown_regex(marked.InlineLexer.rules.zulip, "autolink");
 
-    update_linkifier_rules(linkifiers);
-
     // Tell our fenced code preprocessor how to insert arbitrary
     // HTML into the output. This generated HTML is safe to not escape
     fenced_code.set_stash_func((html) => marked.stashHtml(html, true));
@@ -547,7 +448,6 @@ export function initialize(linkifiers, helper_config) {
         unicodeEmojiHandler: handleUnicodeEmoji,
         streamHandler: handleStream,
         streamTopicHandler: handleStreamTopic,
-        linkifierHandler: handleLinkifier,
         texHandler: handleTex,
         timestampHandler: handleTimestamp,
         renderer: r,
